Build LLVM's Gold plugin and copy it to the host's target library folder. Additionally, build a private copy of ld.gold & ar when targeting PNaCl.
This adds a submodule for NaCl's binutils. Originally, I had intended to only
use it for it's headers. However, I soon hit runtime C++ ABI incombatiblities
between LLVMgold.so && le32-nacl-ld.gold owing to our LLVM's use of C++11 vs
their LLVM's disuse of C++11 (though that is a slight amount of speculation). In
the end, I chose to build our own copies of ld.gold && ar for use with PNaCl
targets, thereby avoiding use of any part of the PNaCl toolchain outside of
native deps & tests.

Note that the code supporting use of this new linker is still forthcoming.