Add a new attribute: `main_link_name`.
This new attribute is used to rename the link name of the main function
generated by rustc to whatever the user needs. For example, on PNaCl
the rust tool drivers need to have a main named `nacl_main` to work with the
extra libs used to wrap things that aren't otherwise supported in the NaCl
sandbox.